ADDED: A method to calculate maximum float value along with test cases

+++ src/main/MaximumNumber.java
@@ -3,14 +3,15 @@ package main;
 public class MaximumNumber {
     public static void main(String[] args) {
         MaximumNumber maximumNumber = new MaximumNumber();
-        System.out.println("The maximum value is "+maximumNumber.maximumTest(1,2,3));
+        System.out.println("The maximum value is " + maximumNumber.maximumTest(1f, 2f, 3f));
     }
-    public Integer maximumTest(Integer x, Integer y, Integer z){
-        Integer max=x;
-        if(y.compareTo(max)>0)
-            max=y;
-        if(z.compareTo(max)>0)
-            max=z;
+
+    public Float maximumTest(Float x, Float y, Float z) {
+        Float max = x;
+        if (y.compareTo(max) > 0)
+            max = y;
+        if (z.compareTo(max) > 0)
+            max = z;
         return max;
     }
 }
