+++ src/main/TestDemo.java
@@ -1,8 +1,6 @@
 package main;
-import junit.framework.Assert;
-import org.junit.Test;
 
-import java.util.Optional;
+import org.junit.Test;
 
 import static org.junit.Assert.*;
 
@@ -10,15 +8,17 @@ public class TestDemo {
     MaximumNumber maximumNumber = new MaximumNumber();
 
     @Test
-    public void testCase1(){
-        assertEquals(Integer.valueOf(10),maximumNumber.maximumTest(10,5,1));
+    public void testCase1() {
+        assertEquals(Float.valueOf(10), maximumNumber.maximumTest(10f, 5f, 1f));
     }
+
     @Test
-    public void testCase2(){
-        assertEquals(Integer.valueOf(10),maximumNumber.maximumTest(5,10,1));
+    public void testCase2() {
+        assertEquals(Float.valueOf(10), maximumNumber.maximumTest(5f, 10f, 1f));
     }
+
     @Test
-    public void testCase3(){
-        assertEquals(Integer.valueOf(10),maximumNumber.maximumTest(1,5,10));
+    public void testCase3() {
+        assertEquals(Float.valueOf(10), maximumNumber.maximumTest(1f, 5f, 10f));
     }
 }
